#72: Extract managing test assets to AssetTestRule
